# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AppState": {
        "properties": {
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/LabelProfile"
            },
            "title": "Profiles",
            "type": "array"
          },
          "queue_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Queue Name",
            "type": "string"
          },
          "selected_profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Profile Id"
          },
          "stock_is_continuous": {
            "default": false,
            "title": "Stock Is Continuous",
            "type": "boolean"
          },
          "stock_length_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 29.0,
            "title": "Stock Length Mm"
          },
          "stock_width_mm": {
            "default": 62.0,
            "exclusiveMinimum": 0.0,
            "title": "Stock Width Mm",
            "type": "number"
          }
        },
        "required": [
          "queue_name"
        ],
        "title": "AppState",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LabelBorderInput": {
        "properties": {
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "inset_mm": {
            "default": 1.0,
            "minimum": 0.0,
            "title": "Inset Mm",
            "type": "number"
          },
          "radius_mm": {
            "default": 1.5,
            "minimum": 0.0,
            "title": "Radius Mm",
            "type": "number"
          },
          "thickness_mm": {
            "default": 0.5,
            "exclusiveMinimum": 0.0,
            "title": "Thickness Mm",
            "type": "number"
          }
        },
        "title": "LabelBorderInput",
        "type": "object"
      },
      "LabelProfile": {
        "properties": {
          "border": {
            "$ref": "#/components/schemas/LabelBorderInput"
          },
          "cut_every": {
            "default": 1,
            "minimum": 1.0,
            "title": "Cut Every",
            "type": "integer"
          },
          "height_mm": {
            "exclusiveMinimum": 0.0,
            "title": "Height Mm",
            "type": "number"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "orientation": {
            "default": "portrait",
            "title": "Orientation",
            "type": "string"
          },
          "quality": {
            "default": "BrQuality",
            "title": "Quality",
            "type": "string"
          },
          "quantity": {
            "default": 1,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/LabelRowInput"
            },
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          },
          "width_mm": {
            "exclusiveMinimum": 0.0,
            "title": "Width Mm",
            "type": "number"
          }
        },
        "required": [
          "name",
          "rows",
          "width_mm",
          "height_mm",
          "id"
        ],
        "title": "LabelProfile",
        "type": "object"
      },
      "LabelProfileInput": {
        "properties": {
          "border": {
            "$ref": "#/components/schemas/LabelBorderInput"
          },
          "cut_every": {
            "default": 1,
            "minimum": 1.0,
            "title": "Cut Every",
            "type": "integer"
          },
          "height_mm": {
            "exclusiveMinimum": 0.0,
            "title": "Height Mm",
            "type": "number"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "orientation": {
            "default": "portrait",
            "title": "Orientation",
            "type": "string"
          },
          "quality": {
            "default": "BrQuality",
            "title": "Quality",
            "type": "string"
          },
          "quantity": {
            "default": 1,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/LabelRowInput"
            },
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          },
          "width_mm": {
            "exclusiveMinimum": 0.0,
            "title": "Width Mm",
            "type": "number"
          }
        },
        "required": [
          "name",
          "rows",
          "width_mm",
          "height_mm"
        ],
        "title": "LabelProfileInput",
        "type": "object"
      },
      "LabelRowInput": {
        "properties": {
          "alignment": {
            "default": "left",
            "title": "Alignment",
            "type": "string"
          },
          "bold": {
            "default": false,
            "title": "Bold",
            "type": "boolean"
          },
          "italic": {
            "default": false,
            "title": "Italic",
            "type": "boolean"
          },
          "level": {
            "default": "normal",
            "title": "Level",
            "type": "string"
          },
          "text": {
            "default": "",
            "maxLength": 300,
            "title": "Text",
            "type": "string"
          }
        },
        "title": "LabelRowInput",
        "type": "object"
      },
      "PrintJobResult": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "queue": {
            "title": "Queue",
            "type": "string"
          },
          "stderr": {
            "default": "",
            "title": "Stderr",
            "type": "string"
          },
          "stdout": {
            "default": "",
            "title": "Stdout",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "queue"
        ],
        "title": "PrintJobResult",
        "type": "object"
      },
      "QueueConfig": {
        "properties": {
          "queue_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Queue Name",
            "type": "string"
          },
          "stock_is_continuous": {
            "default": false,
            "title": "Stock Is Continuous",
            "type": "boolean"
          },
          "stock_length_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 29.0,
            "title": "Stock Length Mm"
          },
          "stock_width_mm": {
            "default": 62.0,
            "exclusiveMinimum": 0.0,
            "title": "Stock Width Mm",
            "type": "number"
          }
        },
        "required": [
          "queue_name"
        ],
        "title": "QueueConfig",
        "type": "object"
      },
      "QueueState": {
        "properties": {
          "default_queue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Queue"
          },
          "queue_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Queue Name",
            "type": "string"
          },
          "queues": {
            "items": {
              "type": "string"
            },
            "title": "Queues",
            "type": "array"
          },
          "stock_is_continuous": {
            "default": false,
            "title": "Stock Is Continuous",
            "type": "boolean"
          },
          "stock_length_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 29.0,
            "title": "Stock Length Mm"
          },
          "stock_width_mm": {
            "default": 62.0,
            "exclusiveMinimum": 0.0,
            "title": "Stock Width Mm",
            "type": "number"
          }
        },
        "required": [
          "queue_name"
        ],
        "title": "QueueState",
        "type": "object"
      },
      "QueueStatus": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "is_detected": {
            "default": false,
            "title": "Is Detected",
            "type": "boolean"
          },
          "is_online": {
            "default": false,
            "title": "Is Online",
            "type": "boolean"
          },
          "job_ids": {
            "items": {
              "type": "string"
            },
            "title": "Job Ids",
            "type": "array"
          },
          "queue_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Queue Name",
            "type": "string"
          },
          "queued_jobs": {
            "default": 0,
            "minimum": 0.0,
            "title": "Queued Jobs",
            "type": "integer"
          },
          "status": {
            "default": "unknown",
            "maxLength": 100,
            "minLength": 1,
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "queue_name"
        ],
        "title": "QueueStatus",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/config": {
      "get": {
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueState"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config"
      },
      "post": {
        "operationId": "save_config_api_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Config"
      }
    },
    "/api/config/defaults": {
      "get": {
        "operationId": "get_config_defaults_api_config_defaults_get",
        "parameters": [
          {
            "in": "query",
            "name": "queue_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Queue Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Config Defaults Api Config Defaults Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config Defaults"
      }
    },
    "/api/config/options": {
      "get": {
        "operationId": "get_config_options_api_config_options_get",
        "parameters": [
          {
            "in": "query",
            "name": "queue_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Queue Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "type": "object"
                  },
                  "title": "Response Get Config Options Api Config Options Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config Options"
      }
    },
    "/api/profiles": {
      "post": {
        "operationId": "create_profile_endpoint_api_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile Endpoint"
      }
    },
    "/api/profiles/{profile_id}": {
      "delete": {
        "operationId": "delete_profile_endpoint_api_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile Endpoint"
      },
      "put": {
        "operationId": "update_profile_endpoint_api_profiles__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile Endpoint"
      }
    },
    "/api/profiles/{profile_id}/select": {
      "post": {
        "operationId": "select_profile_endpoint_api_profiles__profile_id__select_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Profile Endpoint"
      }
    },
    "/api/queue-status": {
      "get": {
        "operationId": "get_queue_status_api_queue_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "queue_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Queue Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Queue Status"
      }
    },
    "/api/state": {
      "get": {
        "operationId": "get_state_api_state_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppState"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get State"
      }
    },
    "/labels.pdf": {
      "post": {
        "operationId": "generate_label_pdf_labels_pdf_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Label Pdf"
      }
    },
    "/print": {
      "post": {
        "operationId": "print_label_print_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintJobResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Print Label"
      }
    }
  }
}